+++ app/src/js/components/Collections/granules.js
@@ -158,18 +158,15 @@ const CollectionGranules = ({
           filterAction={filterGranules}
           filterClear={clearGranulesFilter}
         >
+          <Search
+            action={searchGranules}
+            clear={clearGranulesSearch}
+            label="Search"
+            labelKey="granuleId"
+            placeholder="Granule ID"
+            searchKey="granules"
+          />
           <ListFilters>
-            <Search
-              action={searchGranules}
-              clear={clearGranulesSearch}
-              inputProps={{
-                className: 'search search--large',
-              }}
-              label="Search"
-              labelKey="granuleId"
-              placeholder="Granule ID"
-              searchKey="granules"
-            />
             {view === 'all' && (
               <Dropdown
                 options={statusOptions}
@@ -190,7 +187,8 @@ const CollectionGranules = ({
               paramKey="provider"
               label="Provider"
               inputProps={{
-                placeholder: 'All'
+                placeholder: 'All',
+                className: 'dropdown--medium',
               }}
             />
           </ListFilters>